% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmend.R
\name{convertnet}
\alias{convertnet}
\title{Export network to Cytoscape}
\usage{
convertnet(res, fdrth = 0.05)
}
\arguments{
\item{res}{output ggmend}

\item{fdrth}{Significant level of FDR. default = 0.05}
}
\description{
This function exports a network in edge and node list files in a format suitable for importing to \href{https://cran.r-project.org/web/packages/visNetwork/vignettes/Introduction-to-visNetwork.html}{visNetwork}
}
\examples{
data(data_ggmend)
res <- ggmend(data_ggmend$egenes, data_ggmend$cissnps,
              effectth = 0, corth = 0, standardize = TRUE,
              maxperms = 1000, threads = 1,seed = 1)
netdata <- convertnet(res)
plotnet(netdata$nodes, netdata$edges)
}
