% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmend.R
\name{ggmend}
\alias{ggmend}
\title{GGmend (Gene-on-Gene effect estimator using Mendelian randomization)}
\usage{
ggmend(ys, xs, effectth = 0, corth = 0, standardize = TRUE,
  maxperms = 1000, threads = 1, seed = 0)
}
\arguments{
\item{ys}{A \emph{m} by \emph{n} matrix, where \emph{m} is the number of egenes, and \emph{n} is the number of individuals}

\item{xs}{A \emph{m} by \emph{n} matrix, where \emph{m} is the number of \emph{cis}-acting snps, and \emph{n} is the number of individuals}

\item{effectth}{effect size filtering threshold. default = 0}

\item{corth}{correlation filtering threshold. default = 0}

\item{standardize}{logical value. If false ys and xs transform the genotype data and the expression data for each gene to have mean 0 and variance 1}

\item{maxperms}{Number of maximum permutation. default = 1000}

\item{threads}{The number of cores to use. default = 1}

\item{seed}{set.seed}
}
\description{
ggmend accurately identifies gene-on-gene regulatory effects utilizing a quartet of two genes and their two \emph{cis}-eQTLs
}
\details{
\emph{ys} (egenes) and \emph{xs} (cissnps) require the same dimension. Parallelization (\emph{threads} > 1) works only on *nix (Linux, Unix such as macOS) system. please check \emph{.Platform$OS.type} function.
}
\examples{
data(data_ggmend)
res <- ggmend(data_ggmend$egenes, data_ggmend$cissnps,
              effectth = 0, corth = 0, standardize = TRUE,
              maxperms = 1000, threads = 1,seed = 1)
netdata <- convertnet(res, fdrth = 0.05)
plotnet(netdata$nodes, netdata$edges)
}
\keyword{ggmend}
