% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmend.R
\name{plotnet}
\alias{plotnet}
\title{Network visualization for GGMend}
\usage{
plotnet(nodes, edges)
}
\arguments{
\item{nodes}{See online documentation \href{https://cran.r-project.org/web/packages/visNetwork/vignettes/Introduction-to-visNetwork.html}{visNetwork}}

\item{edges}{See online documentation \href{https://cran.r-project.org/web/packages/visNetwork/vignettes/Introduction-to-visNetwork.html}{visNetwork}}
}
\description{
Network visualization using \href{https://cran.r-project.org/web/packages/visNetwork/vignettes/Introduction-to-visNetwork.html}{visNetwork}
}
\examples{
data(data_ggmend)
res <- ggmend(data_ggmend$egenes, data_ggmend$cissnps,
              effectth = 0, corth = 0, standardize = TRUE,
              maxperms = 1000, threads = 1,seed = 1)
netdata <- convertnet(res, fdrth = 0.05)
plotnet(netdata$nodes, netdata$edges)
}
